Added debug option
You can now use the debug method in background.js to show a modal in the options menu. Helpful for testing in builds.
Also started implementing the shortcut creation, still having trouble with command line arguments, but I did manage to figure out how to parse them. However, the entire main process broke when I tried to do it, so more work needs to be done.
I think I'll be okay to release this in it's current state. Bugs and issues are inevitable, and if I keep putting off until features x, y, and z are ready, I'll keep adding features to the list and never release the damn thing.